[SES-47] Removed old components
