--- src/stories/components/custom-multi-select-2/custom-multi-select-2.scss
@@ -1,3 +1,0 @@
-.MuiListItemText-primary {
-  font-size: 14px;
-}
