+++ src/stories/components/custom-multi-select/custom-multi-select.scss
@@ -1,5 +1,3 @@
-.CustomOutline {
-  .MuiSelect-select.MuiSelect-outlined.MuiSelect-multiple.MuiOutlinedInput-input.MuiInputBase-input {
-    padding: 11px 32px 11px 14px;
-  }
+.MuiListItemText-primary {
+  font-size: 14px;
 }
