--- src/stories/components/search-input/search-input.scss
@@ -1,6 +1,0 @@
-.CustomSearchInput {
-  .MuiOutlinedInput-input.MuiInputBase-input.MuiInputBase-inputAdornedEnd {
-    padding: 11px 14px;
-    font-size: 14px;
-  }
-}
