+++ src/stories/components/search-input/search-input.tsx
@@ -1,60 +1,57 @@
-import React, { useState } from 'react';
-import { FormControl, InputAdornment, InputLabel, OutlinedInput } from '@mui/material';
-import { Search } from '@mui/icons-material';
-import './search-input.scss';
+import React, { CSSProperties } from 'react';
+import styled from '@emotion/styled';
+import Magnifier from '../svg/magnifier';
 
 interface SearchInputProps {
-  label: string,
   value?: string,
   placeholder: string,
   onChange?: (text: string) => void
+  style?: CSSProperties,
 }
 
 export const SearchInput = (props: SearchInputProps) => {
-  const [value, setValue] = useState(props.value || '');
-  const [oldTimeout, setOldTimeout] = useState<NodeJS.Timeout>();
-
   const handleChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    setValue(event.target.value);
-    oldTimeout && clearTimeout(oldTimeout);
-    setOldTimeout(setTimeout(() => {
-      props.onChange && props.onChange(event.target.value);
-    }, 600));
+    props.onChange && props.onChange(event.target.value);
   };
 
-  const [active, setActive] = useState(false);
-
-  return <FormControl
-    sx={{
-      m: '10px 8px',
-      width: '300px',
-      background: 'white'
-    }}
-    variant="outlined"
-    onFocus={() => setActive(true)}
-    onBlur={() => setActive(false)}
-  >
-    <InputLabel
-      sx={{
-        fontSize: '14px',
-        lineHeight: '16px',
-        top: '-2px'
-      }}
-      htmlFor="outlined-adornment-password">
-      {active || value ? props.label : props.placeholder}
-    </InputLabel>
-    <OutlinedInput
-      id="outlined-adornment-password"
-      type={'text'}
-      value={value}
-      onChange={handleChange}
-      className={'CustomSearchInput'}
-      endAdornment={
-        <InputAdornment position="end">
-          <Search sx={{ m: 1 }}/>
-        </InputAdornment>
-      }
-      label={props.label}
-    />
-  </FormControl>;
+  return <Container style={props.style}>
+    <InputWrapper>
+      <Input onChange={handleChange} placeholder={props.placeholder} value={props.value}/>
+      <IconWrapper><Magnifier/></IconWrapper>
+    </InputWrapper>
+  </Container>;
 };
+
+const Container = styled.div({
+  display: 'flex',
+});
+
+const InputWrapper = styled.div({
+  position: 'relative',
+  display: 'flex',
+  alignItems: 'center'
+});
+
+const Input = styled.input({
+  fontFamily: 'SF Pro Text, sans-serif',
+  fontStyle: 'normal',
+  fontWeight: 500,
+  fontSize: '14px',
+  flex: 1,
+  outline: 'none',
+  width: '320px',
+  height: '48px',
+  border: '1px solid #D4D9E1',
+  borderRadius: '22px',
+  padding: '15px 45px 15px 16px',
+  boxSizing: 'border-box',
+  transition: 'all .3s ease',
+  '&:focus': {
+    borderColor: '#25273D'
+  }
+});
+
+const IconWrapper = styled.div({
+  position: 'absolute',
+  right: '22px'
+});
